+++ lib/ts/recipe/thirdparty/providers/utils.ts
@@ -1,7 +1,103 @@
+import * as jose from "jose";
 import { ProviderConfigForClientType } from "../types";
 import { getOIDCDiscoveryInfo } from "../../../thirdpartyUtils";
 import NormalisedURLDomain from "../../../normalisedURLDomain";
 import NormalisedURLPath from "../../../normalisedURLPath";
+import { logDebugMessage } from "../../../logger";
+import { doFetch } from "../../../utils";
+
+export async function doGetRequest(
+    url: string,
+    queryParams?: { [key: string]: string },
+    headers?: { [key: string]: string }
+): Promise<{
+    jsonResponse: Record<string, any> | undefined;
+    status: number;
+    stringResponse: string;
+}> {
+    logDebugMessage(
+        `GET request to ${url}, with query params ${JSON.stringify(queryParams)} and headers ${JSON.stringify(headers)}`
+    );
+    if (headers?.["Accept"] === undefined) {
+        headers = {
+            ...headers,
+            Accept: "application/json", // few providers like github don't send back json response by default
+        };
+    }
+    const finalURL = new URL(url);
+    finalURL.search = new URLSearchParams(queryParams).toString();
+    let response = await doFetch(finalURL.toString(), {
+        headers: headers,
+    });
+
+    const stringResponse = await response.text();
+    let jsonResponse: Record<string, any> | undefined = undefined;
+
+    logDebugMessage(`Received response with status ${response.status} and body ${stringResponse}`);
+
+    if (response.status < 400) {
+        jsonResponse = JSON.parse(stringResponse);
+    }
+
+    return {
+        stringResponse,
+        status: response.status,
+        jsonResponse,
+    };
+}
+
+export async function doPostRequest(
+    url: string,
+    params: { [key: string]: any },
+    headers?: { [key: string]: string }
+): Promise<{
+    jsonResponse: Record<string, any> | undefined;
+    status: number;
+    stringResponse: string;
+}> {
+    if (headers === undefined) {
+        headers = {};
+    }
+
+    headers["Content-Type"] = "application/x-www-form-urlencoded";
+    headers["Accept"] = "application/json"; // few providers like github don't send back json response by default
+
+    logDebugMessage(
+        `POST request to ${url}, with params ${JSON.stringify(params)} and headers ${JSON.stringify(headers)}`
+    );
+
+    const body = new URLSearchParams(params).toString();
+    let response = await doFetch(url, {
+        method: "POST",
+        body,
+        headers,
+    });
+
+    const stringResponse = await response.text();
+    let jsonResponse: Record<string, any> | undefined = undefined;
+
+    logDebugMessage(`Received response with status ${response.status} and body ${stringResponse}`);
+
+    if (response.status < 400) {
+        jsonResponse = JSON.parse(stringResponse);
+    }
+
+    return {
+        stringResponse,
+        status: response.status,
+        jsonResponse,
+    };
+}
+
+export async function verifyIdTokenFromJWKSEndpointAndGetPayload(
+    idToken: string,
+    jwks: jose.JWTVerifyGetKey,
+    otherOptions: jose.JWTVerifyOptions
+): Promise<any> {
+    const { payload } = await jose.jwtVerify(idToken, jwks, otherOptions);
+
+    return payload;
+}
 
 export async function discoverOIDCEndpoints(config: ProviderConfigForClientType): Promise<void> {
     if (config.oidcDiscoveryEndpoint !== undefined) {