+++ lib/ts/supertokens.ts
@@ -20,6 +20,7 @@ import {
     normaliseHttpMethod,
     sendNon200ResponseWithMessage,
     getRidFromHeader,
+    isTestEnv,
 } from "./utils";
 import { Querier } from "./querier";
 import RecipeModule from "./recipeModule";
@@ -150,7 +151,7 @@ export default class SuperTokens {
         if (!oauth2Found) {
             this.recipeModules.push(OAuth2ProviderRecipe.init()(this.appInfo, this.isInServerlessEnv));
         }
-        this.telemetryEnabled = config.telemetry === undefined ? process.env.TEST_MODE !== "testing" : config.telemetry;
+        this.telemetryEnabled = config.telemetry === undefined ? !isTestEnv() : config.telemetry;
     }
 
     static init(config: TypeInput) {
@@ -161,7 +162,7 @@ export default class SuperTokens {
     }
 
     static reset() {
-        if (process.env.TEST_MODE !== "testing") {
+        if (!isTestEnv()) {
             throw new Error("calling testing function in non testing env");
         }
         Querier.reset();